chore(reorganize): restructuring the parser wrapper (#476)
