+++ servers/parser-graphql-wrapper/src/apollo/typeDefs.ts
@@ -3,5 +3,5 @@ import fs from 'fs';
 import { gql } from 'graphql-tag';
 
 export const typeDefs = gql(
-  fs.readFileSync(path.join(__dirname, '..', 'schema.graphql')).toString(),
+  fs.readFileSync(path.join(__dirname, '../..', 'schema.graphql')).toString(),
 );
