--- servers/parser-graphql-wrapper/src/logger.ts
@@ -1,2 +1,0 @@
-import { Logger, setLogger } from '@pocket-tools/ts-logger';
-export const serverLogger: Logger = setLogger();
