+++ servers/parser-graphql-wrapper/src/test/queries/marticleParser.integration.ts
@@ -7,15 +7,15 @@
  */
 
 import nock, { cleanAll } from 'nock';
-import { getRedis } from '../cache';
-import { VideoType } from '../__generated__/resolvers-types';
-import { startServer } from '../server';
-import { ParserAPI } from '../datasources/parserApi';
+import { getRedis } from '../../cache';
+import { VideoType } from '../../__generated__/resolvers-types';
+import { startServer } from '../../apollo/server';
+import { ParserAPI } from '../../datasources/parserApi';
 import { ApolloServer } from '@apollo/server';
 import request from 'supertest';
 import { print } from 'graphql';
 import { gql } from 'graphql-tag';
-import { IContext } from '../context';
+import { IContext } from '../../apollo/context';
 import { Application } from 'express';
 
 function makeResponseForParserTextEndpoint(options: {
