+++ servers/parser-graphql-wrapper/src/test/queries/readerViewResult.integration.ts
@@ -1,15 +1,18 @@
 import nock, { cleanAll } from 'nock';
-import { getRedis } from '../cache';
-import { startServer } from '../server';
+import { getRedis } from '../../cache';
+import { startServer } from '../../apollo/server';
 import { ApolloServer } from '@apollo/server';
 import request from 'supertest';
 import { print } from 'graphql';
 import { gql } from 'graphql-tag';
-import { IContext } from '../context';
+import { IContext } from '../../apollo/context';
 import { Application } from 'express';
 import { DataSource } from 'typeorm';
-import { getConnection, getSharedUrlsConnection } from '../database/mysql';
-import { ItemResolver } from '../entities/ItemResolver';
+import {
+  getConnection,
+  getSharedUrlsConnection,
+} from '../../datasources/mysql';
+import { ItemResolver } from '../../entities/ItemResolver';
 import { IntMask } from '@pocket-tools/int-mask';
 import * as ogs from 'open-graph-scraper';
 jest.mock('open-graph-scraper');
