+++ servers/parser-graphql-wrapper/tsconfig.json
@@ -1,9 +1,9 @@
 {
-  "extends": "tsconfig/cdktf.json",
+  "extends": "tsconfig/graphql.json",
   "compilerOptions": {
     "outDir": "dist",
     "rootDir": "src"
   },
-  "exclude": ["node_modules"],
-  "include": ["src/**/*.ts", "src/config"]
+  "exclude": ["node_modules/", "dist/"],
+  "include": ["src/**/*.ts"]
 }
